Add files via upload
Algorithmic Trading Strategy for Amazon stocks

The Momentum Strategy, one of the most common trading strategies, is used because it's assumed that stocks have momentum or upward or downward trends, that you can detect and exploit. 
Specifically, Dual Moving Average Crossover is used to know when a short-term average crosses a long-term average. This signal is used to identify that momentum is shifting in the direction of the short-term average. A buy signal is generated when the short-term average crosses the long-term average and rises above it, while a sell signal is triggered by a short-term average crossing long-term average and falling below it.